chore: Python 3.10+ type modernization + add diagnostic scripts
- Replace Optional[X] / Union[X, None] with X | None (PEP 604)
- Replace Generic[T] class syntax with PEP 695 class Foo[T]
- Replace try/except pass with contextlib.suppress
- Add diagnostic/investigation scripts from TV-parity session
  (scripts/_debug_*, scripts/_check_*, scripts/_diagnose_*,
   scripts/_compare_*, scripts/_verify_all_metrics.py, etc.)
- Add scripts/run_uvicorn.bat

All changes are non-functional style/type modernization.

+++ backend/backtesting/currency_utils.py
@@ -241,7 +241,7 @@ def price_to_ticks(price: float, symbol: str) -> int:
     Returns:
         Number of ticks
     """
-    return int(round(price / get_min_tick(symbol)))
+    return round(price / get_min_tick(symbol))
 
 
 # Update cache with actual rates (would use API in production)
